+++ src/components/chat/chat-input.js
@@ -8,12 +8,14 @@ import { Label } from "../ui/label";
 import { Textarea } from "src/components/ui/textarea";
 import { utils } from "@ricky0123/vad-react";
 import {useSocket} from "../providers/sock-provider";
+import { useChatData } from "../hooks/use-chat-data";
 
 
 const ChatInput = ({ channelId, userInfo, setData }) => {
     const [enabled, setEnabled] = useState(false); // 채팅번역 기능
     const { socket, isConnected } = useSocket();
     const { onOpen } = useModal();
+    const { updateData } = useChatData()
 
     const sendMessageRef = useRef(null);
     /***
@@ -31,8 +33,6 @@ const ChatInput = ({ channelId, userInfo, setData }) => {
     useEffect(() => {
         if (!channelId || !isConnected || !userInfo) return;
 
-        const updateData = (chatData) => setData((prev) => [...prev, chatData]);
-
         const subscription = socket.subscribe(
             `/topic/msg/${channelId}`,
             async (msg) => {
@@ -49,29 +49,8 @@ const ChatInput = ({ channelId, userInfo, setData }) => {
 
                     const { data } = await axios(axiosConfig);
                     result = data;
-                } else if (
-                    result.chat.type === "AUDIO" &&
-                    result.chat.sender !== parseInt(userInfo?.sub)
-                ) {
-                    console.log("AUDIO 시작");
-                    const apiUrl = `http://localhost:8000/api/v1/audio/audio/${userInfo?.national_language}`;
-                    const axiosConfig = {
-                        url: apiUrl,
-                        method: "POST",
-                        responseType: "blob",
-                        data: { ...result.chat },
-                    };
-
-                    axios(axiosConfig).then((response) => {
-                        const url = window.URL.createObjectURL(
-                            new Blob([response.data])
-                        );
-                        const audio = new Audio(url);
-                        audio.play();
-                    });
-
-                    return;
-                }
+                } 
+                
                 updateData(result.chat);
             }
         );
